+++ drivers/src/lp_filter.c
@@ -5,8 +5,8 @@ void filter_init(LowPassFilter *filter, int alpha, int y0) {
     filter->y_prev = y0;
 }
 
-int apply_filter(LowPassFilter *filter, int x) {
-    int y = filter->alpha * x + (1 - filter->alpha) * filter->y_prev;
+signed long apply_filter(LowPassFilter *filter, signed long x) {
+    signed long y = filter->alpha * x + (1 - filter->alpha) * filter->y_prev;
     filter->y_prev = y;
     return y;
 }
